style: redesign site
Inspired by: https://github.com/owickstrom/the-monospace-web

I really liked the design and structure of this css demo.  The
foundational metric in the mono lib is the `line-height` CSS property.
As such, we created a variable and then built all vertical spacing
around `line-height`: All margins, font sizes, padding, etc, are all
based on it.